requirements-base.txt
biopython
pyteomics
requests
requests-cache